add check so gpt.js is only loaded once
This fixes an issue where repeated calls to .dfp would result in
multiple copies of the doubleclick javascript.

This doesn't just keep track of its own state because it's possible for
gpt.js to get injected some external means.

Uses a kinda crappy selector, `src*=` because it could be http... it
could be https... and there could be GET parameters. But it's only run
once. After the first time, dfpIsLoaded becomes truthy and the selector
isn't evaluated again.

The script blocking detection appears to still work.