pass the slotRenderedEvent to afterEachAdLoaded
I'm using jquery.dfp.js and davidjbradshaw/iframe-resizer. In order for the iframe-resizer to work accurately the iframe widths or heights have to be fluid if you want the iframe to expand horizontally or vertically, respectively.  Given there are cases where there may be more than one configured dimension on the ad unit, it is amazingly helpful to have the rendered ad ```size``` as passed back from the ```slotRenderEnded``` event.  In this way, I can resize the containing div to the base rendered ad size then set the iframe to have a fluid dimension to allow for expandability.